:wrench: Change configuration for mypy to use django-stubs :rocket:
We need to use mypy_django_plugin so that mypy would know where should
it looking for stub file. This plugin required only 1 thing which is we
need to tell where is the django_settings_module is for the project

Another configuration I add is [mypy-*.migrations.*] which we gonna
ignore all error produce from the migrations file. Declaring it like
this will apply to the app we will add in the future as well without
re-configuration mypy again

The only downside for using mypy-django-plugins is that we need to run
it at the root of django project, So that mypy would be able to collect
the module for each app (related to this issue typeddjango/django-stubs#133)

Lastly, we need to add import os in production settings since mypy
wouldn't know the import * format